// src/app/models/selected-model/selected-model.ts
export interface SelectedModel {
    selectedCode: string;
    selectedModelDescription: string;
    selectedColor: string;
    selectedColorPrice: number;
    selectedConfigDescription: string;
    selectedModelPrice: number;
    selectedRange: number;
    selectedSpeed: number;
    selectedTowHitch: number;
    selectedYoke: number;
    selectedImageUrl: string
}